Move routing related components to routes directory
