+++ app/components/routes/LoginPage.tsx
@@ -7,15 +7,15 @@ import { withRouter } from 'react-router-dom';
 import styled from 'styled-components';
 import type { ClientConfig } from 'backend/routes/Config.types';
 
-import { login } from '../actions/manager/auth';
-import type { ReduxState } from '../reducers';
-import renderMultilineText from '../utils/shared/renderMultilineText';
-import SplashLoadingScreen from '../utils/SplashLoadingScreen';
-import StageUtils from '../utils/stageUtils';
-import LargeLogo from './banner/LargeLogo';
-import LogoLabel from './banner/LogoLabel';
-import { Button, Form, FullScreenSegment, Input, Message, Popup } from './basic';
-import type { ReduxThunkDispatch } from '../configureStore';
+import { login } from '../../actions/manager/auth';
+import type { ReduxState } from '../../reducers';
+import renderMultilineText from '../../utils/shared/renderMultilineText';
+import SplashLoadingScreen from '../../utils/SplashLoadingScreen';
+import StageUtils from '../../utils/stageUtils';
+import LargeLogo from '../banner/LargeLogo';
+import LogoLabel from '../banner/LogoLabel';
+import { Button, Form, FullScreenSegment, Input, Message, Popup } from '../basic';
+import type { ReduxThunkDispatch } from '../../configureStore';
 
 export interface LoginPageProps {
     isLoggingIn: boolean;
