+++ app/components/routes/Routes.tsx
@@ -4,12 +4,12 @@ import { useSelector } from 'react-redux';
 import { Route, Redirect, Switch } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
 
-import Consts from '../utils/consts';
+import Consts from '../../utils/consts';
 import LogoPage from './LogoPage';
 import LoginPage from './LoginPage';
 import ExternalRedirect from './ExternalRedirect';
 import AuthRoutes from './AuthRoutes';
-import type { ReduxState } from '../reducers';
+import type { ReduxState } from '../../reducers';
 import ExternalLogin from './ExternalLogin';
 
 const Routes: FunctionComponent = () => {
